Feedback from comments
